Adds textAlign() to Text component.
Crafty doesn't have a built-in way to center text. This extends the
Text component to add CSS text alignment to DOM text elements. It's
impossible to add text alignment to Canvas Text objects right now
because Crafty's Text draw function is not exposed. The canvas needs
its textAlign property set in the middle of the draw operation.